requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML and NLP
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
datasets==2.14.6

# Vector database
pinecone-client==2.2.4

# RSS and web scraping
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8

# Data processing
pandas==2.1.3
numpy==1.24.4

# Environment and configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
